refactor: separate app and test folders
