+++ src/app/lib/is-resource-valid.ts
